add abstract startupcheck
