+++ confignode/src/main/java/org/apache/iotdb/confignode/service/ConfigNodeCommandLine.java
@@ -20,11 +20,9 @@ package org.apache.iotdb.confignode.service;
 
 import org.apache.iotdb.common.rpc.thrift.TConfigNodeLocation;
 import org.apache.iotdb.commons.ServerCommandLine;
-import org.apache.iotdb.commons.conf.IoTDBConstant;
 import org.apache.iotdb.commons.exception.BadNodeUrlException;
 import org.apache.iotdb.commons.exception.ConfigurationException;
 import org.apache.iotdb.commons.exception.StartupException;
-import org.apache.iotdb.commons.service.StartupChecks;
 import org.apache.iotdb.confignode.conf.ConfigNodeRemoveCheck;
 import org.apache.iotdb.confignode.conf.ConfigNodeStartupCheck;
 
@@ -70,8 +68,8 @@ public class ConfigNodeCommandLine extends ServerCommandLine {
     if (MODE_START.equals(mode)) {
       try {
         // Startup environment check
-        StartupChecks checks = new StartupChecks(IoTDBConstant.CN_ROLE).withDefaultTest();
-        checks.verify();
+        //        StartupChecks checks = new StartupChecks(IoTDBConstant.CN_ROLE).withDefaultTest();
+        //        checks.verify();
         // Do ConfigNode startup checks
         ConfigNodeStartupCheck.getInstance().startUpCheck();
       } catch (StartupException | ConfigurationException | IOException e) {
